Complete cleanup and unification of classes
