--- setup.cfg
@@ -1,2 +1,0 @@
-[metadata]
-description-file=README.md
