+++ setup.py
@@ -5,26 +5,26 @@ with open("README.md", "r") as infile:
 
 setuptools.setup(
     name="sololearnlib",
-    version="1.0.4",
+    version="1.0.5",
     author="OR!ON",
-    author_email='orionconner00@gmail.com', 
+    author_email="orionconner00@gmail.com", 
     license="MIT",
     description="Retrieve public data from 'sololearn.com'.",
     long_description=readme_file,
-    url='https://github.com/Or-i0n/pkg_sololearnlib',
+    url="https://github.com/Or-i0n/sololearnlib",
     packages=setuptools.find_packages(),
     install_requires=[           
-          'bs4',
+          "bs4",
       ],
   classifiers=[
-    'Development Status :: 3 - Alpha',      
-    'Intended Audience :: Developers',      
-    'Topic :: Software Development :: Build Tools',
-    'License :: OSI Approved :: MIT License',
-    'Programming Language :: Python :: 3.5',
-    'Programming Language :: Python :: 3.6',
-    'Programming Language :: Python :: 3.7',
-    'Programming Language :: Python :: 3.8',
+    "Development Status :: 3 - Alpha",      
+    "Intended Audience :: Developers",      
+    "Topic :: Software Development :: Build Tools",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.5",
+    "Programming Language :: Python :: 3.6",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
 
   ],
   python_requires=">=3.5"
